moved styles to styles folder for workout page
